[Move `@kbn/config-schema` to server] `stack_connectors`
