+++ x-pack/plugins/stack_connectors/common/bedrock/types.ts
@@ -5,35 +5,19 @@
  * 2.0.
  */
 
-import { TypeOf } from '@kbn/config-schema';
-import {
-  ConfigSchema,
-  DashboardActionParamsSchema,
-  DashboardActionResponseSchema,
-  SecretsSchema,
-  RunActionParamsSchema,
-  RunActionResponseSchema,
-  InvokeAIActionParamsSchema,
-  InvokeAIActionResponseSchema,
-  InvokeAIRawActionParamsSchema,
-  InvokeAIRawActionResponseSchema,
-  StreamingResponseSchema,
-  RunApiLatestResponseSchema,
-  BedrockMessageSchema,
-  BedrockToolChoiceSchema,
-} from './schema';
-
-export type Config = TypeOf<typeof ConfigSchema>;
-export type Secrets = TypeOf<typeof SecretsSchema>;
-export type RunActionParams = TypeOf<typeof RunActionParamsSchema>;
-export type InvokeAIActionParams = TypeOf<typeof InvokeAIActionParamsSchema>;
-export type InvokeAIActionResponse = TypeOf<typeof InvokeAIActionResponseSchema>;
-export type InvokeAIRawActionParams = TypeOf<typeof InvokeAIRawActionParamsSchema>;
-export type InvokeAIRawActionResponse = TypeOf<typeof InvokeAIRawActionResponseSchema>;
-export type RunApiLatestResponse = TypeOf<typeof RunApiLatestResponseSchema>;
-export type RunActionResponse = TypeOf<typeof RunActionResponseSchema>;
-export type StreamingResponse = TypeOf<typeof StreamingResponseSchema>;
-export type DashboardActionParams = TypeOf<typeof DashboardActionParamsSchema>;
-export type DashboardActionResponse = TypeOf<typeof DashboardActionResponseSchema>;
-export type BedRockMessage = TypeOf<typeof BedrockMessageSchema>;
-export type BedrockToolChoice = TypeOf<typeof BedrockToolChoiceSchema>;
+export type {
+  Config,
+  Secrets,
+  RunActionParams,
+  InvokeAIActionParams,
+  InvokeAIActionResponse,
+  InvokeAIRawActionParams,
+  InvokeAIRawActionResponse,
+  RunApiLatestResponse,
+  RunActionResponse,
+  StreamingResponse,
+  DashboardActionParams,
+  DashboardActionResponse,
+  BedRockMessage,
+  BedrockToolChoice,
+} from '../../server/connector_types/bedrock/schema';
