+++ x-pack/plugins/stack_connectors/common/slack_api/types.ts
@@ -6,29 +6,28 @@
  */
 
 import type { ActionType as ConnectorType } from '@kbn/actions-plugin/server/types';
-import { TypeOf } from '@kbn/config-schema';
 import type { ActionTypeExecutorOptions as ConnectorTypeExecutorOptions } from '@kbn/actions-plugin/server/types';
 import type { ActionTypeExecutorResult as ConnectorTypeExecutorResult } from '@kbn/actions-plugin/server/types';
-import {
-  PostMessageParamsSchema,
-  PostMessageSubActionParamsSchema,
-  PostBlockkitSubActionParamsSchema,
-  PostBlockkitParamsSchema,
-  SlackApiSecretsSchema,
-  SlackApiParamsSchema,
-  SlackApiConfigSchema,
-  ValidChannelIdSubActionParamsSchema,
-} from './schema';
+import type {
+  PostBlockkitSubActionParams,
+  PostMessageSubActionParams,
+  SlackApiConfig,
+  SlackApiParams,
+  SlackApiSecrets,
+} from '../../server/connector_types/slack_api/schema';
 
-export type SlackApiSecrets = TypeOf<typeof SlackApiSecretsSchema>;
-export type SlackApiConfig = TypeOf<typeof SlackApiConfigSchema>;
+export type {
+  SlackApiSecrets,
+  SlackApiConfig,
+  PostMessageParams,
+  PostMessageSubActionParams,
+  PostBlockkitSubActionParams,
+  PostBlockkitParams,
+  ValidChannelIdSubActionParams,
+  SlackApiParams,
+  SlackApiActionParams,
+} from '../../server/connector_types/slack_api/schema';
 
-export type PostMessageParams = TypeOf<typeof PostMessageParamsSchema>;
-export type PostMessageSubActionParams = TypeOf<typeof PostMessageSubActionParamsSchema>;
-export type PostBlockkitSubActionParams = TypeOf<typeof PostBlockkitSubActionParamsSchema>;
-export type PostBlockkitParams = TypeOf<typeof PostBlockkitParamsSchema>;
-export type ValidChannelIdSubActionParams = TypeOf<typeof ValidChannelIdSubActionParamsSchema>;
-export type SlackApiParams = TypeOf<typeof SlackApiParamsSchema>;
 export type SlackApiConnectorType = ConnectorType<
   SlackApiConfig,
   SlackApiSecrets,
@@ -47,8 +46,6 @@ export type SlackExecutorOptions = ConnectorTypeExecutorOptions<
   SlackApiSecrets,
   SlackApiParams
 >;
-
-export type SlackApiActionParams = TypeOf<typeof SlackApiParamsSchema>;
 
 export interface SlackAPiResponse {
   ok: boolean;
