+++ x-pack/plugins/stack_connectors/server/connector_types/bedrock/index.ts
@@ -19,7 +19,7 @@ import { urlAllowListValidator } from '@kbn/actions-plugin/server';
 import { ValidatorServices } from '@kbn/actions-plugin/server/types';
 import { assertURL } from '@kbn/actions-plugin/server/sub_action_framework/helpers/validators';
 import { BEDROCK_CONNECTOR_ID, BEDROCK_TITLE } from '../../../common/bedrock/constants';
-import { ConfigSchema, SecretsSchema } from '../../../common/bedrock/schema';
+import { ConfigSchema, SecretsSchema } from './schema';
 import { Config, Secrets } from '../../../common/bedrock/types';
 import { BedrockConnector } from './bedrock';
 import { renderParameterTemplates } from './render';
