+++ x-pack/plugins/stack_connectors/server/connector_types/gemini/index.ts
@@ -19,7 +19,7 @@ import { ValidatorServices } from '@kbn/actions-plugin/server/types';
 import { assertURL } from '@kbn/actions-plugin/server/sub_action_framework/helpers/validators';
 import { GenerativeAIForObservabilityConnectorFeatureId } from '@kbn/actions-plugin/common/connector_feature_config';
 import { GEMINI_CONNECTOR_ID, GEMINI_TITLE } from '../../../common/gemini/constants';
-import { ConfigSchema, SecretsSchema } from '../../../common/gemini/schema';
+import { ConfigSchema, SecretsSchema } from './schema';
 import { Config, Secrets } from '../../../common/gemini/types';
 import { GeminiConnector } from './gemini';
 import { renderParameterTemplates } from './render';
