+++ x-pack/plugins/stack_connectors/server/connector_types/openai/openai.test.ts
@@ -16,7 +16,7 @@ import {
 } from '../../../common/openai/constants';
 import { loggingSystemMock } from '@kbn/core-logging-server-mocks';
 import { actionsMock } from '@kbn/actions-plugin/server/mocks';
-import { RunActionResponseSchema, StreamingResponseSchema } from '../../../common/openai/schema';
+import { RunActionResponseSchema, StreamingResponseSchema } from './schema';
 import { initDashboard } from '../lib/gen_ai/create_gen_ai_dashboard';
 import { PassThrough, Transform } from 'stream';
 import { ConnectorUsageCollector } from '@kbn/actions-plugin/server/types';
