+++ x-pack/plugins/stack_connectors/server/connector_types/slack_api/schema.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { schema } from '@kbn/config-schema';
+import { TypeOf, schema } from '@kbn/config-schema';
 
 export const SlackApiSecretsSchema = schema.object({
   token: schema.string({ minLength: 1 }),
@@ -71,3 +71,14 @@ export const SlackApiParamsSchema = schema.oneOf([
   PostMessageParamsSchema,
   PostBlockkitParamsSchema,
 ]);
+
+export type SlackApiSecrets = TypeOf<typeof SlackApiSecretsSchema>;
+export type SlackApiConfig = TypeOf<typeof SlackApiConfigSchema>;
+
+export type PostMessageParams = TypeOf<typeof PostMessageParamsSchema>;
+export type PostMessageSubActionParams = TypeOf<typeof PostMessageSubActionParamsSchema>;
+export type PostBlockkitSubActionParams = TypeOf<typeof PostBlockkitSubActionParamsSchema>;
+export type PostBlockkitParams = TypeOf<typeof PostBlockkitParamsSchema>;
+export type ValidChannelIdSubActionParams = TypeOf<typeof ValidChannelIdSubActionParamsSchema>;
+export type SlackApiParams = TypeOf<typeof SlackApiParamsSchema>;
+export type SlackApiActionParams = TypeOf<typeof SlackApiParamsSchema>;
