+++ x-pack/plugins/stack_connectors/server/connector_types/thehive/index.ts
@@ -16,11 +16,7 @@ import {
 } from '@kbn/actions-plugin/common/types';
 import { urlAllowListValidator } from '@kbn/actions-plugin/server';
 import { TheHiveConnector } from './thehive';
-import {
-  TheHiveConfigSchema,
-  TheHiveSecretsSchema,
-  PushToServiceIncidentSchema,
-} from '../../../common/thehive/schema';
+import { TheHiveConfigSchema, TheHiveSecretsSchema, PushToServiceIncidentSchema } from './schema';
 import { THEHIVE_CONNECTOR_ID, THEHIVE_TITLE } from '../../../common/thehive/constants';
 import { TheHiveConfig, TheHiveSecrets } from '../../../common/thehive/types';
