+++ x-pack/plugins/stack_connectors/server/connector_types/thehive/thehive.ts
@@ -16,7 +16,7 @@ import {
   TheHiveAddCommentResponseSchema,
   TheHiveCreateAlertResponseSchema,
   ExecutorSubActionCreateAlertParamsSchema,
-} from '../../../common/thehive/schema';
+} from './schema';
 import type {
   TheHiveConfig,
   TheHiveSecrets,
