+++ x-pack/plugins/stack_connectors/server/connector_types/tines/index.ts
@@ -12,8 +12,7 @@ import {
 import { SecurityConnectorFeatureId } from '@kbn/actions-plugin/common';
 import { urlAllowListValidator } from '@kbn/actions-plugin/server';
 import { TINES_CONNECTOR_ID, TINES_TITLE } from '../../../common/tines/constants';
-import { TinesConfigSchema, TinesSecretsSchema } from '../../../common/tines/schema';
-import { TinesConfig, TinesSecrets } from '../../../common/tines/types';
+import { TinesConfig, TinesSecrets, TinesConfigSchema, TinesSecretsSchema } from './schema';
 import { TinesConnector } from './tines';
 import { renderParameterTemplates } from './render';
