+++ x-pack/plugins/stack_connectors/server/connector_types/webhook/types.ts
@@ -11,7 +11,7 @@ import type {
   ActionTypeExecutorOptions as ConnectorTypeExecutorOptions,
 } from '@kbn/actions-plugin/server/types';
 import { ParamsSchema, ConfigSchema } from './schema';
-import { SecretConfigurationSchema } from '../../../common/auth/schema';
+import { SecretConfigurationSchema } from '../auth/schema';
 
 export type WebhookConnectorType = ConnectorType<
   ConnectorTypeConfigType,
